Moved all the context-related codes back to contexts/modalContext.js
