--- frontend/src/components/modalContext/ModalContext.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-const ModalContext = React.createContext();
-
-export default ModalContext;
